renamed classes related to setup wizard
