+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/setupwindow/AbstractStepRepresentation.java
@@ -37,15 +37,15 @@ import java.net.URL;
 // TODO: Fix this class using CSS and containers instead of static sizes
 abstract class AbstractStepRepresentation {
     private final Message<?> messageWaitingForResponse;
-    private final SetupWindowJavaFXImplementation parent;
+    private final SetupUiJavaFXImplementation parent;
     private Button nextButton;
 
-    AbstractStepRepresentation(SetupWindowJavaFXImplementation parent, Message<?> messageWaitingForResponse) {
+    AbstractStepRepresentation(SetupUiJavaFXImplementation parent, Message<?> messageWaitingForResponse) {
         this.parent = parent;
         this.messageWaitingForResponse = messageWaitingForResponse;
     }
 
-    protected SetupWindowJavaFXImplementation getParent() {
+    protected SetupUiJavaFXImplementation getParent() {
         return this.parent;
     }
