+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/setupwindow/StepRepresentationBrowse.java
@@ -38,7 +38,7 @@ public class StepRepresentationBrowse extends AbstractStepRepresentationWithHead
     private List<String> extensions;
     private File selectedFile;
 
-    public StepRepresentationBrowse(SetupWindowJavaFXImplementation parent, Message<String> message,
+    public StepRepresentationBrowse(SetupUiJavaFXImplementation parent, Message<String> message,
                                     String textToShow, File browseDirectory, List<String> extensions) {
         super(parent, message);
         this.browseDirectory = browseDirectory;
