+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/setupwindow/StepRepresentationLicence.java
@@ -28,7 +28,7 @@ public class StepRepresentationLicence extends StepRepresentationMessage {
     private final String licenceText;
     private boolean isAgree;
 
-    public StepRepresentationLicence(SetupWindowJavaFXImplementation parent, Message messageWaitingForResponse, String textToShow,
+    public StepRepresentationLicence(SetupUiJavaFXImplementation parent, Message messageWaitingForResponse, String textToShow,
                                      String licenceText) {
         super(parent, messageWaitingForResponse, textToShow);
