+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/views/setupwindow/StepRepresentationTextBox.java
@@ -26,7 +26,7 @@ public class StepRepresentationTextBox extends StepRepresentationMessage {
     private final Message<String> messageWaitingForResponse;
     private TextField textField;
 
-    public StepRepresentationTextBox(SetupWindowJavaFXImplementation parent,
+    public StepRepresentationTextBox(SetupUiJavaFXImplementation parent,
                                      Message<String> messageWaitingForResponse,
                                      String textToShow,
                                      String defaultValue) {
