+++ phoenicis-scripts/src/main/java/org/phoenicis/scripts/ui/UiQuestionFactory.java
@@ -18,21 +18,10 @@
 
 package org.phoenicis.scripts.ui;
 
-import java.util.function.Consumer;
-import java.util.function.Supplier;
+public interface UiQuestionFactory {
+    void create(String questionText, Runnable yesCallback, Runnable noCallback);
 
-public interface UIMessageSender {
-    <R> R run(Supplier<R> function);
-
-    default <R> R runAndWait(Consumer<Message<R>> function) {
-        final Message<R> message = new Message<>();
-        run(() -> {
-            function.accept(message);
-            return null;
-        });
-        message.block();
-
-        return message.get();
+    default void create(String questionText, Runnable yesCallback) {
+        create(questionText, yesCallback, () -> {});
     }
-
 }
